Revert file renaming, minor changes to docstrings
